Move SetupUseSnippetSwitch to setup.
