+++ assets/js/modules/analytics-4/components/common/index.js
@@ -18,5 +18,4 @@
 
 export { default as PropertySelect } from './PropertySelect';
 export { default as SettingsUseSnippetSwitch } from './SettingsUseSnippetSwitch';
-export { default as SetupUseSnippetSwitch } from './SetupUseSnippetSwitch';
 export { default as UseSnippetSwitch } from './UseSnippetSwitch';
