+++ assets/js/modules/analytics/components/setup/SetupFormGA4Transitional.js
@@ -44,9 +44,9 @@ import {
 	ProfileNameTextField,
 	GA4PropertyNotice,
 	ExistingGTMPropertyNotice,
-	SetupUseSnippetSwitch as SetupUseSnippetSwitchUA,
 } from '../common';
-import { SetupUseSnippetSwitch as SetupUseSnippetSwitchGA4 } from '../../../analytics-4/components/common';
+import { SetupUseSnippetSwitch as SetupUseSnippetSwitchUA } from '../setup';
+import { SetupUseSnippetSwitch as SetupUseSnippetSwitchGA4 } from '../../../analytics-4/components/setup';
 const { useSelect } = Data;
 
 export default function SetupFormGA4Transitional() {
