+++ assets/js/modules/tagmanager/components/setup/index.js
@@ -19,3 +19,4 @@
 export { default as SetupErrorNotice } from './SetupErrorNotice';
 export { default as SetupForm } from './SetupForm';
 export { default as SetupMain } from './SetupMain';
+export { default as SetupUseSnippetSwitch } from './SetupUseSnippetSwitch';
